requirements.txt
configobj~=5.0.8
Flask~=3.0.0
peewee==3.17.0
Flask-Limiter[redis]~=3.5.0
pycryptodome~=3.19.0